Struct rspsps::Parser

pub struct Parser {
    // some fields omitted
}

This is the safe abstraction over a spsps parser.

Methods

impl Parser

fn new() -> Parser

Create a new Parser with zero configuration. Name will use the default from libspsps, and will read from stdin.

fn from_file(name: Option<&str>, path: &Path) -> Parser

Create a Parser that will read from the file given in path. If None is given for name the default spsps parser name will be used.

fn consume(&self) -> String

Consume and return the next character in the stream.

fn consume_n(&self, n: usize)

Consume and discard the next n characters from the stream.

fn consume_whitespace(&self)

Consume and discard all whitespace. When this method returns the next character is the first non-whitespace character, or the end of file has been reached.

fn at_eof(&self) -> bool

Determine if the end of file has been consumed.

fn get_loc(&self) -> Loc

Get the current location in the stream. This is the location of the next character to be read, unless the end of stream has been reached.

fn peek(&self) -> String

Peek and return the next character in the stream. The character is not consumed.

fn peek_n(&self, n: usize) -> String

Peek ahead at the next few characters in the stream, and return them. The The number of characters must be below the lookahead limit. End of file causes the returned string to be populated by EOF characters.

fn peek_str(&self, needle: &str) -> bool

Peek ahead and determine if the next characters in the stream are the given characters, in sequence. That is, the given string must be the next thing in the stream.

fn peek_str_and_consume(&self, needle: &str) -> bool

Peek ahead at the next few characters and if they are a given string, then consume them. Otherwise leave the stream unchanged.